find the even and odd
An Even or Odd number program is a basic Java application used to check whether a given number is even or odd. A number is called even if it is divisible by 2 and gives no remainder, and odd if it gives a remainder when divided by 2. This program uses simple Java concepts such as the Scanner class, if–else statement, and modulus operator. It is useful for beginners to understand decision-making in Java programming